pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"


[project]
name = "pyfleascope"
version = "0.0.3"
authors = [
  { name="D. Freiermuth", email="author@example.com" },
]
description = "A python package for instrumenting the FleaScope"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyserial",
    "pandas",
    "pyudev",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/daniel-freiermuth/pyfleascope"
Issues = "https://github.com/daniel-freiermuth/pyfleascope/issues"
